Updated compile to implementation, etc.
In the app build.gradle, the compile and testCompile calls have been replaced with implementation and testImplementation respectively.

In the lib build.gradle, compile and testCompile have been replaced with api and testImplentation respectively.

This is to prevent any issues at the end of 2018 when compile and testCompile are no longer supported by Google/Gradle.